Refactor .d.ts files and imports
