+++ javascript/components/index.js
@@ -1,0 +1,63 @@
+import AbstractLayer from './AbstractLayer';
+import CircleLayer from './CircleLayer';
+import Images from './Images';
+import NativeUserLocation from './NativeUserLocation';
+import SkyLayer from './SkyLayer';
+import AbstractSource from './AbstractSource';
+import FillExtrusionLayer from './FillExtrusionLayer';
+import Light from './Light';
+import PointAnnotation from './PointAnnotation';
+import Style from './Style';
+import Annotation from './Annotation';
+import FillLayer from './FillLayer';
+import LineLayer from './LineLayer';
+import RasterDemSource from './RasterDemSource';
+import SymbolLayer from './SymbolLayer';
+import BackgroundLayer from './BackgroundLayer';
+import HeadingIndicator from './HeadingIndicator';
+import MapView from './MapView';
+import RasterLayer from './RasterLayer';
+import Terrain from './Terrain';
+import Callout from './Callout';
+import HeatmapLayer from './HeatmapLayer';
+import MarkerView from './MarkerView';
+import RasterSource from './RasterSource';
+import UserLocation from './UserLocation';
+import ImageSource from './ImageSource';
+import NativeBridgeComponent from './NativeBridgeComponent';
+import ShapeSource from './ShapeSource';
+import VectorSource from './VectorSource';
+import Camera from './Camera';
+
+export {
+  AbstractLayer,
+  CircleLayer,
+  Images,
+  NativeUserLocation,
+  SkyLayer,
+  AbstractSource,
+  FillExtrusionLayer,
+  Light,
+  PointAnnotation,
+  Style,
+  Annotation,
+  FillLayer,
+  LineLayer,
+  RasterDemSource,
+  SymbolLayer,
+  BackgroundLayer,
+  HeadingIndicator,
+  MapView,
+  RasterLayer,
+  Terrain,
+  Callout,
+  HeatmapLayer,
+  MarkerView,
+  RasterSource,
+  UserLocation,
+  ImageSource,
+  NativeBridgeComponent,
+  ShapeSource,
+  VectorSource,
+  Camera,
+};
