+++ javascript/modules/location/locationManager.d.ts
@@ -1,0 +1,4 @@
+export class LocationManager extends Component {
+  start(displacement?: number): void;
+  stop(): void;
+}
