+++ javascript/modules/snapshot/snapshotManager.d.ts
@@ -1,0 +1,3 @@
+export class SnapshotManager {
+  static takeSnap(options: SnapshotOptions): Promise<string>;
+}
