+++ javascript/utils/Logger.d.ts
@@ -1,0 +1,4 @@
+export class Logger {
+  public static setLogCallback: (cb: LogCallback) => boolean;
+  public static setLogLevel: (level: LogLevel) => void;
+}
