+++ javascript/utils/animated/index.js
@@ -1,0 +1,15 @@
+import Animated from './Animated';
+import AnimatedPoint from './AnimatedPoint';
+import AnimatedShape from './AnimatedShape';
+import AnimatedCoordinatesArray from './AnimatedCoordinatesArray';
+import AnimatedExtractCoordinateFromArray from './AnimatedExtractCoordinateFromArray';
+import AnimatedRouteCoordinatesArray from './AnimatedRouteCoordinatesArray';
+
+export {
+  Animated,
+  AnimatedPoint,
+  AnimatedShape,
+  AnimatedCoordinatesArray,
+  AnimatedExtractCoordinateFromArray,
+  AnimatedRouteCoordinatesArray,
+};
