+++ javascript/utils/index.d.ts
@@ -1,0 +1,1 @@
+export function isAndroid(): boolean;
