+++ package.json
@@ -6,7 +6,8 @@
     "access": "public"
   },
   "author": "React Native Mapbox Team",
-  "main": "index.ts",
+  "main": "./lib/index.js",
+  "types": "./types",
   "keywords": [
     "gl",
     "ios",
