+++ scripts/autogenerate.js
@@ -416,11 +416,6 @@ async function generate() {
       output: path.join(IOS_OUTPUT_PATH, 'RCTMGLStyle.h'),
       only: 'gl',
     },
-    // TODO: This doesn't seem right?
-    // {
-    //   input: path.join(TMPL_PATH, 'definitions.d.ts.ejs'),
-    //   output: path.join(IOS_OUTPUT_PATH, 'definitions.d.ts'),
-    // },
     {
       input: path.join(TMPL_PATH, 'RCTMGLStyle.m.ejs'),
       output: path.join(IOS_OUTPUT_PATH, 'RCTMGLStyle.m'),
@@ -476,11 +471,15 @@ async function generate() {
   execSync('yarn build:plugin');
   outputPaths.push('plugin/build');
 
+  // compile TS and JS
+  console.log('Compiling ./lib directory');
+  execSync('npx tsc; cp -R ./assets ./lib/assets');
+
   // autogenerate docs
   const docBuilder = new DocJSONBuilder(layers);
-  const markdownBuilder = new MarkdownBuilder();
   await docBuilder.generate();
-  await markdownBuilder.generate();
+  const markdownBuilder = new MarkdownBuilder();
+  markdownBuilder.generate();
 
   // Check if any generated files changed
   try {
